Refactor the world
